godon - evacuate root dag definition to controller
Might be best handled by breeder creater routines in future.
