+++ controller/breeder_create.py
@@ -6,6 +6,31 @@ from .meta_data_db import META_DB_CONFIG
 
 import datetime
 
+def determine_config_shard(run_id=None,
+                           target_id=None,
+                           config=None,
+                           targets_count=0,
+                           parallel_runs_count=0):
+
+    config_result = copy.deepcopy(config)
+    settings_space = config_result.get('settings').get('sysctl')
+
+    for setting in settings_space.items():
+        upper = setting.get('constraints').get('upper')
+        lower = setting.get('constraints').get('lower')
+
+        delta = abs(upper - lower)
+
+        shard_size = delta / targets_count * parallel_runs_count
+
+        setting['constraints']['lower'] = lower + shard_size * (run_id + target_id)
+
+    config_result['settings']['sysctl'] = settings_space
+
+    return config_result
+
+
+
 
 def main(request_data):
 
@@ -61,6 +86,21 @@ def main(request_data):
                                                   creation_ts=datetime.datetime.now(),
                                                   meta_state=breeder_config)
     archive.archive_db.execute(db_info=db_config, query=__query)
+
+    # TODO - adjust breeder flows initialization to wmill way
+
+#    for target in targets:
+#        hash_suffix = hashlib.sha256(str.encode(target.get('address'))).hexdigest()[0:6]
+#
+#        for run_id in range(0, parallel_runs):
+#            dag_id = f'{dag_name}_{run_id}'
+#            if not is_cooperative:
+#                config = determine_config_shard()
+#            globals()[f'{dag_id}_optimization_{hash_suffix}'] = create_optimization_dag(f'{dag_id}_optimization_{hash_suffix}', config, run_id, hash_suffix)
+#            globals()[f'{dag_id}_target_{hash_suffix}'] = create_target_interaction_dag(f'{dag_id}_target_interaction_{hash_suffix}', config, target, hash_suffix)
+#
+#        target_id += 1
+
 
     return { "result": "SUCCESS", "breeder_id": uuid }
